+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### unreleased
 - **add:** npm keywords
+- **fix:** replace chrome/firefox with phantomjs
 
 ### [v0.1.0] - 2016-06-11
 - **fix:** move minified version